small fixes

+++ NightlifeEvents.Api/Models/Event.cs
@@ -7,6 +7,6 @@ namespace NightlifeEvents.Models
         public string City { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public DateTime Date { get; set; }
-        public int TicketPrice { get; set; }
+        public decimal TicketPrice { get; set; }
     }
 }
